+++ src/sklearn_evaluation/plot/__init__.py
@@ -11,9 +11,12 @@ from .grid_search import grid_search
 from .validation_curve import validation_curve
 from .learning_curve import learning_curve
 from .metrics import metrics_at_thresholds
+from .regression import residuals, prediction_error
 from .clustering import elbow_curve, elbow_curve_from_results
-from .calibration import calibration_curve
 from .target_analysis import target_plot
+from .calibration import calibration_curve, scores_distribution
+from .classification_report import classification_report, ClassificationReport
+
 
 __all__ = [
     'confusion_matrix',
@@ -26,8 +29,13 @@ __all__ = [
     'learning_curve',
     'metrics_at_thresholds',
     'ConfusionMatrix',
+    'residuals',
+    'prediction_error',
     'elbow_curve',
     'elbow_curve_from_results',
     'calibration_curve',
     'target_plot'
+    'scores_distribution',
+    'classification_report',
+    'ClassificationReport',
 ]